// src/payload-generated-schema.ts
export const enum_blogs_status = pgEnum('enum_blogs_status', ['draft', 'published'])
// ...
export const users = pgTable(
  'users',
  {
    id: serial('id').primaryKey(),
    updatedAt: timestamp('updated_at', { mode: 'string', withTimezone: true, precision: 3 })
      .defaultNow()
      .notNull(),
    createdAt: timestamp('created_at', { mode: 'string', withTimezone: true, precision: 3 })
      .defaultNow()
      .notNull(),
    email: varchar('email').notNull(),
    resetPasswordToken: varchar('reset_password_token'),
    resetPasswordExpiration: timestamp('reset_password_expiration', {
      mode: 'string',
      withTimezone: true,
      precision: 3,
    }),
    salt: varchar('salt'),
    hash: varchar('hash'),
    loginAttempts: numeric('login_attempts', { mode: 'number' }).default(0),
    lockUntil: timestamp('lock_until', { mode: 'string', withTimezone: true, precision: 3 }),
  },
  (columns) => [
    index('users_updated_at_idx').on(columns.updatedAt),
    index('users_created_at_idx').on(columns.createdAt),
    uniqueIndex('users_email_idx').on(columns.email),
  ],
)

export const media = pgTable(
  'media',
  {
    id: serial('id').primaryKey(),
    alt: varchar('alt').notNull(),
    updatedAt: timestamp('updated_at', { mode: 'string', withTimezone: true, precision: 3 })
      .defaultNow()
      .notNull(),
    createdAt: timestamp('created_at', { mode: 'string', withTimezone: true, precision: 3 })
      .defaultNow()
      .notNull(),
    url: varchar('url'),
    thumbnailURL: varchar('thumbnail_u_r_l'),
    filename: varchar('filename'),
    mimeType: varchar('mime_type'),
    filesize: numeric('filesize', { mode: 'number' }),
    width: numeric('width', { mode: 'number' }),
    height: numeric('height', { mode: 'number' }),
    focalX: numeric('focal_x', { mode: 'number' }),
    focalY: numeric('focal_y', { mode: 'number' }),
  },
  (columns) => [
    index('media_updated_at_idx').on(columns.updatedAt),
    index('media_created_at_idx').on(columns.createdAt),
    uniqueIndex('media_filename_idx').on(columns.filename),
  ],
)

export const blogs = pgTable(
  'blogs',
  {
    id: serial('id').primaryKey(),
    title: varchar('title'),
    content: jsonb('content'),
    author: integer('author_id').references(() => users.id, {
      onDelete: 'set null',
    }),
    authorSocialLinks_twitter: varchar('author_social_links_twitter'),
    authorSocialLinks_linkedin: varchar('author_social_links_linkedin'),
    authorSocialLinks_github: varchar('author_social_links_github'),
    authorSocialLinks_website: varchar('author_social_links_website'),
    generateSlug: boolean('generate_slug').default(true),
    slug: varchar('slug'),
    bannerImage: integer('banner_image_id').references(() => media.id, {
      onDelete: 'set null',
    }),
    thumbnailImage: integer('thumbnail_image_id').references(() => media.id, {
      onDelete: 'set null',
    }),
    publishDate: timestamp('publish_date', { mode: 'string', withTimezone: true, precision: 3 }),
    updatedAt: timestamp('updated_at', { mode: 'string', withTimezone: true, precision: 3 })
      .defaultNow()
      .notNull(),
    createdAt: timestamp('created_at', { mode: 'string', withTimezone: true, precision: 3 })
      .defaultNow()
      .notNull(),
    _status: enum_blogs_status('_status').default('draft'),
  },
  (columns) => [
    index('blogs_author_idx').on(columns.author),
    uniqueIndex('blogs_slug_idx').on(columns.slug),
    index('blogs_banner_image_idx').on(columns.bannerImage),
    index('blogs_thumbnail_image_idx').on(columns.thumbnailImage),
    index('blogs_updated_at_idx').on(columns.updatedAt),
    index('blogs_created_at_idx').on(columns.createdAt),
    index('blogs__status_idx').on(columns._status),
  ],
)
// ...
export const blog_embeddings = pgTable(
  'blog_embeddings',
  {
    id: serial('id').primaryKey(),
    blog: integer('blog_id')
      .notNull()
      .references(() => blogs.id, {
        onDelete: 'set null',
      }),
    chunkIndex: numeric('chunk_index', { mode: 'number' }).notNull(),
    content: varchar('content').notNull(),
    embedding: jsonb('embedding').notNull(),
    updatedAt: timestamp('updated_at', { mode: 'string', withTimezone: true, precision: 3 })
      .defaultNow()
      .notNull(),
    createdAt: timestamp('created_at', { mode: 'string', withTimezone: true, precision: 3 })
      .defaultNow()
      .notNull(),
  },
  (columns) => [
    index('blog_embeddings_blog_idx').on(columns.blog),
    index('blog_embeddings_updated_at_idx').on(columns.updatedAt),
    index('blog_embeddings_created_at_idx').on(columns.createdAt),
    index('blog_idx').on(columns.blog),
    uniqueIndex('blog_chunkIndex_idx').on(columns.blog, columns.chunkIndex),
  ],
)